requirements.txt
flask
requests
beautifulsoup4
readability-lxml
